Get the ball rolling
This started commit provides the project setup for iOS 9.0. The app delegate was reconstructed to support app running on iOS starting from 9.0.

The project was reorganized so that the structure looks like below:

	Weather/
		Requirements/
			Diagrams/
		Storyboards/
		Modules/
			Main/
		Business logic/
			Views/
			Entities/
			Services/
			Extentions/
		Localization/
		Test plans/

Also the existing project scheme was converted to use test plans.
(Product > Scheme > Convert Scheme to use Test Plans)

Added the app icon with full set of images required.