Make a detailed sketch of [WeatherLayoutView]
This commit provides a [WeatherLayoutView] draft that is sensitive to device orientation for iPhone as well as for iPad.

The draft contains views intended for current weather, forecast hourly, forecast daily, and national alerts.

[WeatherLayoutView] designed with UIStackView objects in a coder way instead of Interface Builder xib manner because it gives more expressive way to have a deal with auto layout constraints for making adjustments that are can be more sensitive both to rotating events and different devices screen sizes.

Also, all calls to NotificationCenter reference was moved to one call in Settings.

Added MainModuleFunctionalTests blank.

Added environment variable with the key <OS_ACTIVITY_MODE> and the value <disable>.