Restructure Weather Module to be on terms with MVVM
This commit provides a dedicated module [Weather] that was structured with the couple of principles: MVVM.

[Weather] module contains the Model [WeatherModel], the View [WeatherLayoutView], the ViewModel [WeatherViewModel].

The Model takes place within the ViewModel. The ViewModel is placed in the View. The ViewModel has the reference to the View.

In the View all references to the business matter data view containers were extracted and placed in the View, so that the ViewModel is able to access directly via the reference to the View.

Also with each business matter data view container [CurrentWeatherView, ForecastDailyView, ForecastHourlyView, WeatherNationalAlertsView] there was allocated space for the appropriate part of the Model.