Make a draft of the weather data flow implementation
This commit provides a draft of the solution for the weather data flow.

Produced the weather data flow solution is based on the idea that [WeatherViewController] should care itself about the weather data so that it's up-to-date.

For bringing this idea to life [WeatherDataAutoUpdater] class was designed to orchestrate the performance of this job with requesting weather data, requesting geo location data, placing and holding weather data.

[WeatherDataAutoUpdater]  instance is placed within [WeatherViewModel] object.

[WeatherDataAutoUpdater]  instance has the reference to the weather data [WeatherDataModel] instance, the instance of [OpenWeatherClient] class, the reference to the [GeoLocationReceiver] instance, and the reference to the [WeatherDataAutoUpdaterDelegate] instance—all of them give necessaries to make the weather data update job done.

[WeatherDataAutoUpdaterDelegate] object is responsible for reloading data on the screen if weather data changed happens.

[WeatherViewController] changed to be responsible for catching such events as the app did become active and the app did enter background that are the time boundaries when the weather data update should be done.