Put auto-updating business logic into [WeatherDataAutoUpdater]
This commit provides the auto-updating weather data mechanism driving by [WeatherDataAutoUpdater].

The app uses an update mechanism based on timer that is specifically configured so that to be fired always in the begining of hour where possible periods look like [1, 5, 10, 20, 30] in minutes.

So, to be on time it calculates a time from now when the timer should be fired at first time in other words it means delay that should be taken after the first update.

The auto-updating mechanism depends on location permission.

While a user not determined about location services, the app makes an update for the default location.

As soon as a user granted location permission, the app gets the current location and immediately makes an update and in case if location services not allowed it makes an update using either the default location or already used location.

Default locations are added for every supported localization, Plus Codes and Coordinate:

[57PR+JX], {52.186648327996274, -1.7075733650987568} for [EN]

[MJ6C+XG], {55.662546456740564, 85.62138369331707} for [RU]

For #OpenWeather service coordinate with two decimal places is enough to provide accurate forecast, for this reason the auto-updating mechanism makes value only of two decimal places in calculations, e.g. comparison.