Cover [WeatherDataAutoUpdater] with unit tests
ADDED:  The set of unit tests for [WeatherDataAutoUpdater] class.

The set of unit tests are conditionally divided into the three parts.

The 1st part of the set makes sure [WeatherDataAutoUpdater] has a right configuration setup.

The 2nd part of the set makes sure [WeatherDataAutoUpdater] demonstrates appropriate internal timer's behaviour.

The 3rd part of the set makes sure [WeatherDataAutoUpdater] does job as intended.

The test approach for testing timer's behaviour is asynchronous and based on expectations.

For all other things the test approach is synchronous and based on call counting.

FIXED: A correction was made to the title of paragraph V in [Requirements.txt].

IMPROVED: All lines of code were formatted in one single way.

ADDED: The unit test for a default location requirement.